feat: update hooks
